dom/xul/test/test_disable_scroll_frame_plain.html

<!DOCTYPE HTML>
<html scrolling="false">
<head>
  <meta charset="utf-8">
  <title>disable scroll frame exposed</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<div style="height: 300vh"></div>
<script>
  // Ensure that disabling the scroll frame isn't exposed to content.
  ok(document.scrollingElement.scrollTopMax > 0, "Scrolling should still work.");
</script>
</body>
</html>
